feat(#105): drop mock data from production paths; route everything through cubit
- Move `Person`, `MediaKind`, `Track`, `MediaSourceLib`, and the hue/format
  helpers to a new `lib/data/frequency_models.dart` (production-safe).
- Move `kPeople` / `kMedia` to `test/fixtures/frequency_mock_data.dart`;
  delete the old `lib/data/frequency_mock_data.dart`. `grep -r
  frequency_mock_data lib/` is now empty (acceptance criterion).
- Drop the `kPeople.first.btDevice` leak in `FrequencyRoomScreen` — the
  local user no longer ships "Sony WH-1000XM5" to real users; the me-row
  reads the audio-output label until a real route is reported.
- Drop the `debugDemoTimers` flag, the demo timers (talk simulation,
  host join request, weak-signal toast), the parallel `_roster`, and the
  `_talkingId` shadow. The cubit's `SessionRoom.roster` is the only
  source of truth for peers; "talking" rides on `ProtocolPeer.talking`.
- Drop the `groupSize` widget param (only used by the demo roster).
- Remove the hard-coded `kMedia[source]!` lookup; `_applyMediaSnapshot`
  now surfaces the host's source verbatim and renders a placeholder
  queue entry. Title/artwork metadata isn't on the wire (#TBD).
- Rewrite `frequency_room_screen_test.dart` to drive the cubit directly
  (real `FrequencySessionCubit` + `applyJoinAccepted`) for any test
  that exercises the roster / drawer; mock-cubit retained for the
  pure-chrome / mute / PTT / transport tests.
- Repoint `frequency_toast_host_test.dart` to the new fixture path.

Closes #105.
